[BUGFIX][saltedpasswords] Failed backend logins not not logged
The logging functions in saltedpasswords are not able to log failed
backend userlogins to TYPO3s syslog, because the inherited
writelog-function gets overridden by a local function. As a result, no
notification e-mail is sent to [warning_email_addr] when a backend user
has multiple failed login attempts. A remote 'attacker' could try to
login to a TYPO3 installations backend numerous of times without being
noticed (no log entry and no warning-email if configured).

Rename the local writelog-function to writeLogMessage and add the
original writelog-functionality, so that failed backend logins are
written to TYPO3s syslog again and all logging/notifications work as
expected.

Change-Id: Ic05b05873e3fd20df675db908ba76b7dd0e5548f
Resolves: #23917
Releases: 4.6, 4.5, 4.4, 4.3
Reviewed-on: http://review.typo3.org/1795
Reviewed-by: Torben Hansen
Tested-by: Torben Hansen
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter