[FEATURE] Add PBKDF2 based password storage to EXT:saltedpasswords
Implement the PBKDF2 password hashing for EXT:saltedpasswords. PBKDF2 is
the NIST recommended way to store passwords in a FIPS compliant way.

The storage format is identical with passlib from Python to enable some
portability as there's no official storage format yet.

The choice of iteration count has been set to a value that matches the
performance characteristics of the Blowfish/Bcrypt hashing.

Resolves: #28230
Releases: master
Change-Id: I87817129b317fd1b4ebf4bae101b66b3fcebb6ce
Reviewed-on: https://review.typo3.org/42342
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>